# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "urtorrentd",
    "description": "HTTP API of urtorrentd, a BitTorrent daemon on the urtorrent library. Units: bytes, bytes per second, seconds, unix seconds; `null` means unknown or unlimited. Errors always have the `ErrorBody` shape with a stable `code`.",
    "contact": {
      "name": "urtorrentd contributors"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/torrents": {
      "get": {
        "tags": [
          "torrents"
        ],
        "summary": "The torrent list, filtered, sorted and paged.",
        "operationId": "list_torrents",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "State filter.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TorrentFilter"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Only this category (`\"\"` = torrents without one).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Only torrents with this tag (`\"\"` = torrents without tags).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hashes",
            "in": "query",
            "description": "Only these info-hashes, `|`-separated.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "private",
            "in": "query",
            "description": "Only private (`true`) or public (`false`) torrents.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort key.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TorrentSort"
            }
          },
          {
            "name": "reverse",
            "in": "query",
            "description": "Sort descending.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Page start.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TorrentSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "torrents"
        ],
        "summary": "Add torrents from magnet links, info-hashes, URLs and base64 `.torrent`\nfiles. Each source succeeds or fails on its own.",
        "operationId": "add_torrents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTorrentsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddTorrentsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/auto-management": {
      "post": {
        "tags": [
          "torrents"
        ],
        "summary": "Automatic management on or off (on moves torrents to their category's path).",
        "operationId": "set_auto_management",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/category": {
      "post": {
        "tags": [
          "torrents"
        ],
        "summary": "Set or clear the category (automatically managed torrents move).",
        "operationId": "set_torrent_category",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/count": {
      "get": {
        "tags": [
          "torrents"
        ],
        "summary": "How many torrents there are.",
        "operationId": "count_torrents",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/delete": {
      "post": {
        "tags": [
          "torrents"
        ],
        "summary": "Remove torrents, optionally with their content.",
        "operationId": "delete_torrents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/force-start": {
      "post": {
        "tags": [
          "torrents"
        ],
        "summary": "Force-start torrents regardless of the queue limits (`value: true`), or\nhand them back to the queue (`false`).",
        "operationId": "force_start_torrents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddTorrentsRequest": {
        "type": "object",
        "description": "Add torrents from URLs, magnet links, info-hashes and `.torrent` files.",
        "properties": {
          "urls": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Magnet links, bare info-hashes, or `http(s)` URLs of `.torrent` files."
          },
          "torrents": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            },
            "description": "`.torrent` files, base64-encoded."
          },
          "options": {
            "$ref": "#/components/schemas/AddOptions",
            "description": "Options applied to every torrent in the request."
          }
        },
        "additionalProperties": false
      },
      "AddTorrentsResponse": {
        "type": "object",
        "description": "The outcome of an add request.",
        "required": [
          "added",
          "failed"
        ],
        "properties": {
          "added": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddedTorrent"
            },
            "description": "Added torrents."
          },
          "failed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddFailure"
            },
            "description": "Sources that failed."
          }
        }
      },
...
      "CategoryRequest": {
        "type": "object",
        "description": "Set or clear the category.",
        "required": [
          "hashes",
          "category"
        ],
        "properties": {
          "hashes": {
            "$ref": "#/components/schemas/Hashes",
            "description": "Target torrents."
          },
          "category": {
            "type": [
              "string",
              "null"
            ],
            "description": "Category name (created if missing); `null` clears it."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Every error response has this body.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "The error."
          }
        }
      },
...
      "ToggleRequest": {
        "type": "object",
        "description": "A bulk action with an on/off value.",
        "required": [
          "hashes",
          "value"
        ],
        "properties": {
          "hashes": {
            "$ref": "#/components/schemas/Hashes",
            "description": "Target torrents."
          },
          "value": {
            "type": "boolean",
            "description": "On or off."
          }
        }
      },
...
      "TorrentFilter": {
        "type": "string",
        "description": "Filters for the torrent list.",
        "enum": [
          "all",
          "downloading",
          "seeding",
          "completed",
          "stopped",
          "running",
          "active",
          "inactive",
          "stalled",
          "stalled_seeding",
          "stalled_downloading",
          "checking",
          "moving",
          "errored"
        ]
      },
...
      "TorrentSort": {
        "type": "string",
        "description": "Sort keys for the torrent list.",
        "enum": [
          "name",
          "size",
          "progress",
          "download_rate",
          "upload_rate",
          "added_on",
          "completed_on",
          "ratio",
          "queue_position",
          "state",
          "uploaded",
          "downloaded",
          "eta",
          "category"
        ]
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key from POST /api/v1/auth/api-key."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "urtorrentd_sid",
        "description": "Set by POST /api/v1/auth/login."
      }
    }
  },
  "security": [
    {
      "session": []
    },
    {
      "api_key": []
    }
  ],
...
}